MAINT: initialize new project structure
This commit initializes a new project structure. A new top-level
packahe, "please" is started, that will be the eventual home for
all the PLEASE source code.

Additionally, to manage the project dependencies, the project will
do a trial using the Python package "Poetry". To begin this
process, the existing project has been initialized with poetry
and the resulting project defintion files, pyproject.toml and
poetry.lock have been created.

Note, in doing the above, the project dependencies are updated to
the latest compatable package versions. Notable changes are numpy
1.20.x, and PyQt6.