ENH: add uview header parser
This commit introduces a minimal parser for UView data file headers.

The parser now reads the file header and extracts three key
parameters required to know how to find the image data in
the rest of the file. These parameters are width, height, and
bits per pixel. With these three pieces of information the image
data can be extracted from the file end.

TODO:
* add more testing for the many errors that may be raised during
parsing.
* add logging to the parsing